refactor: TreeBase abstract class
