+++ yarn-project/merkle-tree/src/indexed_tree/indexed_tree.ts
@@ -80,7 +80,13 @@ export class IndexedTree implements MerkleTree {
     depth: number,
     prefilledSize = 0,
   ): Promise<IndexedTree> {
-    const underlying = await StandardMerkleTree.new(db, hasher, name, depth, hashEncodedTreeValue(initialLeaf, hasher));
+    const underlying = await StandardMerkleTree.new<StandardMerkleTree>(
+      db,
+      hasher,
+      name,
+      depth,
+      hashEncodedTreeValue(initialLeaf, hasher),
+    );
     const tree = new IndexedTree(underlying, hasher, db);
     await tree.init(prefilledSize);
     return tree;
@@ -94,7 +100,12 @@ export class IndexedTree implements MerkleTree {
    * @returns The newly created tree.
    */
   static async fromName(db: LevelUp, hasher: Hasher, name: string): Promise<IndexedTree> {
-    const underlying = await StandardMerkleTree.fromName(db, hasher, name, hashEncodedTreeValue(initialLeaf, hasher));
+    const underlying = await StandardMerkleTree.fromName<StandardMerkleTree>(
+      db,
+      hasher,
+      name,
+      hashEncodedTreeValue(initialLeaf, hasher),
+    );
     const tree = new IndexedTree(underlying, hasher, db);
     await tree.initFromDb();
     return tree;
