+++ yarn-project/merkle-tree/src/standard_tree/standard_tree.test.ts
@@ -5,11 +5,11 @@ import { merkleTreeTestSuite } from '../test/test_suite.js';
 import { StandardMerkleTree } from './standard_tree.js';
 
 const createDb = async (levelUp: levelup.LevelUp, hasher: Hasher, name: string, depth: number) => {
-  return await StandardMerkleTree.new(levelUp, hasher, name, depth);
+  return await StandardMerkleTree.new<StandardMerkleTree>(levelUp, hasher, name, depth);
 };
 
 const createFromName = async (levelUp: levelup.LevelUp, hasher: Hasher, name: string) => {
-  return await StandardMerkleTree.fromName(levelUp, hasher, name);
+  return await StandardMerkleTree.fromName<StandardMerkleTree>(levelUp, hasher, name);
 };
 
 merkleTreeTestSuite('StandardMerkleTree', createDb, createFromName);
